pyproject.toml
[project]
name = "filesystembasedagent"
version = "0.1.0"
description = "A file system based AI agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
fsagent = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
